[FLINK-14516][network] Simplify code after removing non credit-based flow control
